refactor(apps): extract buildSupportedCounterValues helper and use localeCompare sort
- extract shared buildSupportedCounterValues to logic/ (dedup mobile counter-value build logic)
- use localeCompare in all sort comparators (deterministic ordering when names are equal)
