+++ apps/ledger-live-mobile/src/context/LedgerStore.tsx
@@ -4,11 +4,9 @@ import { Store } from "redux";
 import { importPostOnboardingState } from "@ledgerhq/live-common/postOnboarding/actions";
 import { backfillOnboardingDate } from "~/logic/postOnboarding/backfillOnboardingDate";
 import { CounterValuesStateRaw } from "@ledgerhq/live-countervalues/types";
-import {
-  findCryptoCurrencyById,
-  getCryptoCurrencyById,
-} from "@ledgerhq/live-common/currencies/index";
+import { findCryptoCurrencyById } from "@ledgerhq/live-common/currencies/index";
 import { selectSupportedFiats } from "@domain/entity-currency-fiat";
+import { buildSupportedCounterValues } from "~/logic/buildSupportedCounterValues";
 import { InitialQueriesProvider } from "LLM/contexts/InitialQueriesContext";
 import mmkvStorageWrapper from "LLM/storage/mmkvStorageWrapper";
 import { logStartupEvent } from "LLM/utils/logStartupTime";
@@ -325,19 +323,7 @@ async function hydrateCurrencies() {
 
 function updateSupportedCountervalues(store: Store, settingsData: Partial<SettingsState>) {
   const supportedFiats = selectSupportedFiats(store.getState());
-  const bitcoin = getCryptoCurrencyById("bitcoin");
-  const ethereum = getCryptoCurrencyById("ethereum");
-  const possibleIntermediaries = [bitcoin, ethereum];
-
-  const supportedCounterValues = [...supportedFiats, ...possibleIntermediaries]
-    .map(currency => ({
-      value: currency.ticker,
-      ticker: currency.ticker,
-      label: `${currency.name} - ${currency.ticker}`,
-      currency,
-    }))
-    .sort((a, b) => (a.currency.name < b.currency.name ? -1 : 1));
-
+  const supportedCounterValues = buildSupportedCounterValues(supportedFiats);
   store.dispatch(setSupportedCounterValues(supportedCounterValues));
 
   if (
