+++ apps/ledger-live-mobile/src/mvvm/contexts/InitialQueriesContext.tsx
@@ -1,11 +1,11 @@
 import React, { useEffect, useMemo } from "react";
 import { ofacGeoBlockApi } from "@ledgerhq/live-common/api/ofacGeoBlockApi";
-import { getCryptoCurrencyById } from "@ledgerhq/live-common/currencies/index";
 import { useGetSupportedFiatsQuery } from "@domain/api-currency-fiat";
 import { selectSupportedFiats } from "@domain/entity-currency-fiat";
 import { useDispatch, useSelector } from "~/context/hooks";
 import { setSupportedCounterValues } from "~/actions/settings";
 import { selectRemoteFlagsReady } from "@shared/feature-flags";
+import { buildSupportedCounterValues } from "~/logic/buildSupportedCounterValues";
 
 export const InitialQueriesContext = React.createContext({
   ofacResult: { blocked: false, isLoading: true },
@@ -24,17 +24,7 @@ export function InitialQueriesProvider({ children }: React.PropsWithChildren) {
   // Keep supportedCounterValues in sync: re-dispatch whenever the slice updates (fallback → CVS result)
   const fiats = useSelector(selectSupportedFiats);
   useEffect(() => {
-    const bitcoin = getCryptoCurrencyById("bitcoin");
-    const ethereum = getCryptoCurrencyById("ethereum");
-    const supportedCounterValues = [...fiats, bitcoin, ethereum]
-      .map(currency => ({
-        value: currency.ticker,
-        ticker: currency.ticker,
-        label: `${currency.name} - ${currency.ticker}`,
-        currency,
-      }))
-      .sort((a, b) => (a.currency.name < b.currency.name ? -1 : 1));
-    dispatch(setSupportedCounterValues(supportedCounterValues));
+    dispatch(setSupportedCounterValues(buildSupportedCounterValues(fiats)));
   }, [dispatch, fiats]);
   const ofacResult = useMemo(
     () => ({ blocked: ofacQueryResult.data ?? false, isLoading: ofacQueryResult.isLoading }),
